+++ docs/examples/src/ex-00100-simple-plot.ts
@@ -4,10 +4,10 @@
 // skip-doc-start
 import CandyGraph, {
   createDefaultFont,
-  OpaqueLineStrip,
   OrthoAxis,
   LinearScale,
   CartesianCoordinateSystem,
+  LineStrip,
 } from "../../../src";
 
 export default async function SimplePlot(cg: CandyGraph) {
@@ -46,8 +46,8 @@ export default async function SimplePlot(cg: CandyGraph) {
 
   // Render the a line strip representing the x & y data, and axes.
   cg.render(coords, viewport, [
-    new OpaqueLineStrip(cg, xs, ys, {
-      colors: [1, 0.5, 0.0],
+    new LineStrip(cg, xs, ys, {
+      colors: [1, 0.5, 0.0, 1],
       widths: 3 * dpr,
     }),
     new OrthoAxis(cg, coords, "x", font, {