+++ docs/examples/src/ex-00500-relative-time.ts
@@ -7,10 +7,10 @@
 import CandyGraph, {
   createDefaultFont,
   Grid,
-  TransparentLineStrip,
   OrthoAxis,
   LinearScale,
   CartesianCoordinateSystem,
+  LineStrip,
 } from "../../../src";
 
 export default async function RelativeTime(cg: CandyGraph) {
@@ -108,9 +108,9 @@ export default async function RelativeTime(cg: CandyGraph) {
     }
     return {
       timestamp: time,
-      trace: new TransparentLineStrip(cg, xs, ys, {
-        color: [1, 0, 1, 1],
-        width: 3.0 * dpr,
+      trace: new LineStrip(cg, xs, ys, {
+        colors: [1, 0, 1, 1],
+        widths: 3.0 * dpr,
       }),
     };
   }
@@ -134,8 +134,8 @@ export default async function RelativeTime(cg: CandyGraph) {
     for (let i = 0; i < traces.length - 1; i++) {
       const trace = traces[i];
       const age = time - trace.timestamp;
-      trace.trace.color = [0, 0, 0, 0.5 * (1 - age / HISTORY)];
-      trace.trace.width = 1 * dpr;
+      trace.trace.colors.update([0, 0, 0, 0.5 * (1 - age / HISTORY)])
+      trace.trace.widths.update(1 * dpr);
     }
 
     // Update the timeCoords.