+++ docs/examples/src/ex-00900-polar-plot.ts
@@ -5,7 +5,7 @@
 import CandyGraph, {
   createDefaultFont,
   Text,
-  OpaqueLineStrip,
+  LineStrip,
   LineSegments,
   LinearScale,
   PolarCoordinateSystem,
@@ -100,8 +100,8 @@ export default async function PolarPlot(cg: CandyGraph) {
       colors: axisLineColors,
     }),
     ...axisLabels,
-    new OpaqueLineStrip(cg, rhos, thetas, {
-      colors: [1, 0.5, 0],
+    new LineStrip(cg, rhos, thetas, {
+      colors: [1, 0.5, 0, 1],
       widths: 2.5 * dpr,
     }),
   ]);