+++ docs/tutorial/src/doc-00400.ts
@@ -3,7 +3,7 @@ import CandyGraph, {
   createDefaultFont,
   LinearScale,
   CartesianCoordinateSystem,
-  OpaqueLineStrip,
+  LineStrip,
   OrthoAxis,
   Scissor,
 } from "../../../src";
@@ -89,21 +89,21 @@ export default async function doc_00400(cg: CandyGraph) {
       // in our desired color. We'll also apply a scissor in screen space to
       // confine the lines to the plot region. First the `ys0` trace in ~orange:
       new Scissor(cg, 40, 32, viewport.width - 56, viewport.height - 48, true, [
-        new OpaqueLineStrip(cg, xs, ys0, {
-          colors: [0, 0, 0],
+        new LineStrip(cg, xs, ys0, {
+          colors: [0, 0, 0, 1],
           widths: 17,
         }),
-        new OpaqueLineStrip(cg, xs, ys0, {
-          colors: [1, 0.5, 0],
+        new LineStrip(cg, xs, ys0, {
+          colors: [1, 0.5, 0, 1],
           widths: 9,
         }),
         // Then our `ys1` trace in ~blue:
-        new OpaqueLineStrip(cg, xs, ys1, {
-          colors: [0, 0, 0],
+        new LineStrip(cg, xs, ys1, {
+          colors: [0, 0, 0, 1],
           widths: 17,
         }),
-        new OpaqueLineStrip(cg, xs, ys1, {
-          colors: [0, 0.5, 1],
+        new LineStrip(cg, xs, ys1, {
+          colors: [0, 0.5, 1, 1],
           widths: 9,
         }),
       ]),