+++ src/primitives/opaque-line-strip.ts
@@ -81,6 +81,9 @@ export class OpaqueLineStrip extends Primitive {
     options: OpaqueLineStripOptions = {}
   ) {
     super();
+    console.warn(
+      "OpaqueLineStrip is deprecated and will be removed in a future release. Please use LineStrip instead."
+    );
     const opts = { ...DEFAULT_OPTIONS, ...options };
     this.xs = createDataset(cg, xs);
     this.ys = createDataset(cg, ys);